+++ Example/BitmovinConvivaAnalyticsTvOSExample/ViewController.swift
@@ -27,6 +27,11 @@ class ViewController: UIViewController {
         setupBitmovinPlayer()
     }
 
+    deinit {
+        convivaAnalytics?.release()
+        NSLog("[ Example ] ConvivaAnalytics released")
+    }
+
     func setupBitmovinPlayer() {
         // Setup Player
         player = PlayerFactory.create(playerConfig: playerConfig)