+++ Example/Tests/PlayerEventTests.swift
@@ -1,3 +1,4 @@
+// swiftlint:disable file_length
 //
 //  PlayerEventTests.swift
 //  BitmovinConvivaAnalytics_Tests
@@ -12,6 +13,7 @@ import BitmovinPlayer
 import BitmovinConvivaAnalytics
 import ConvivaSDK
 
+// swiftlint:disable:next type_body_length
 class PlayerEventsSpec: QuickSpec {
     // swiftlint:disable:next function_body_length
     override func spec() {
@@ -49,7 +51,7 @@ class PlayerEventsSpec: QuickSpec {
             context("initialize session") {
                 var spy: Spy!
                 beforeEach {
-                    spy = Spy(aClass: CISClientTestDouble.self, functionName: "createSession")
+                    spy = Spy(aClass: CISVideoAnalyticsTestDouble.self, functionName: "reportPlaybackRequested")
                 }
 
                 it("on play") {
@@ -70,7 +72,7 @@ class PlayerEventsSpec: QuickSpec {
 
             context("not initialize session") {
                 it("on ready") {
-                    let spy = Spy(aClass: CISClientTestDouble.self, functionName: "createSession")
+                    let spy = Spy(aClass: CISVideoAnalyticsTestDouble.self, functionName: "reportPlaybackRequested")
                     playerDouble.fakeReadyEvent()
                     expect(spy).toNot(haveBeenCalled())
                 }
@@ -78,7 +80,7 @@ class PlayerEventsSpec: QuickSpec {
 
             context("initialize player state manager") {
                 it("on play") {
-                    let spy = Spy(aClass: PlayerStateManagerTestDouble.self, functionName: "init")
+                    let spy = Spy(aClass: CISVideoAnalyticsTestDouble.self, functionName: "setPlayerInfo")
                     playerDouble.fakePlayEvent()
                     expect(spy).to(haveBeenCalled())
                 }
@@ -86,14 +88,14 @@ class PlayerEventsSpec: QuickSpec {
 
             context("not initialize player state manager") {
                 it("when initializing conviva analytics") {
-                    let spy = Spy(aClass: PlayerStateManagerTestDouble.self, functionName: "init")
+                    let spy = Spy(aClass: CISVideoAnalyticsTestDouble.self, functionName: "setPlayerInfo")
                     expect(spy).toNot(haveBeenCalled())
                 }
             }
 
             context("deinitialize player state manager") {
                 it("on playback finished") {
-                    let spy = Spy(aClass: CISClientTestDouble.self, functionName: "releasePlayerStateManager")
+                    let spy = Spy(aClass: CISVideoAnalyticsTestDouble.self, functionName: "reportPlaybackEnded")
                     playerDouble.fakePlayEvent()
                     playerDouble.fakePlaybackFinishedEvent()
                     expect(spy).to(haveBeenCalled())
@@ -101,9 +103,9 @@ class PlayerEventsSpec: QuickSpec {
             }
 
             context("update playback state") {
-                var spy: Spy!
+                 var spy: Spy!
                 beforeEach {
-                    spy = Spy(aClass: PlayerStateManagerTestDouble.self, functionName: "setPlayerState")
+                    spy = Spy(aClass: CISVideoAnalyticsTestDouble.self, functionName: "reportPlaybackMetric")
                     playerDouble.fakePlayEvent()
                 }
 
@@ -111,7 +113,9 @@ class PlayerEventsSpec: QuickSpec {
                     it("on play") {
                         playerDouble.fakePlayEvent()
                         expect(spy).notTo(
-                            haveBeenCalled(withArgs: ["newState": "\(PlayerState.CONVIVA_PLAYING.rawValue)"])
+                            haveBeenCalled(withArgs: [
+                                CIS_SSDK_PLAYBACK_METRIC_PLAYER_STATE: "\(PlayerState.CONVIVA_PLAYING.rawValue)"
+                            ])
                         )
                     }
                 }
@@ -119,14 +123,18 @@ class PlayerEventsSpec: QuickSpec {
                 it("on playing") {
                     playerDouble.fakePlayingEvent()
                     expect(spy).to(
-                        haveBeenCalled(withArgs: ["newState": "\(PlayerState.CONVIVA_PLAYING.rawValue)"])
+                        haveBeenCalled(withArgs: [
+                            CIS_SSDK_PLAYBACK_METRIC_PLAYER_STATE: "\(PlayerState.CONVIVA_PLAYING.rawValue)"
+                        ])
                     )
                 }
 
                 it("on pause") {
                     playerDouble.fakePauseEvent()
                     expect(spy).to(
-                        haveBeenCalled(withArgs: ["newState": "\(PlayerState.CONVIVA_PAUSED.rawValue)"])
+                        haveBeenCalled(withArgs: [
+                            CIS_SSDK_PLAYBACK_METRIC_PLAYER_STATE: "\(PlayerState.CONVIVA_PAUSED.rawValue)"
+                        ])
                     )
                 }
                 it("on stall started/ Stall Ended wait 0.10 seconds") {
@@ -135,7 +143,9 @@ class PlayerEventsSpec: QuickSpec {
                     playerDouble.fakeStallEndedEvent()
                     DispatchQueue.main.asyncAfter(deadline: .now() + 0.10) {
                         expect(spy).notTo(
-                            haveBeenCalled(withArgs: ["newState": "\(PlayerState.CONVIVA_BUFFERING.rawValue)"])
+                            haveBeenCalled(withArgs: [
+                                CIS_SSDK_PLAYBACK_METRIC_PLAYER_STATE: "\(PlayerState.CONVIVA_BUFFERING.rawValue)"
+                            ])
                         )
                     }
                 }
@@ -144,7 +154,9 @@ class PlayerEventsSpec: QuickSpec {
                     playerDouble.fakeStallStartedEvent()
                     playerDouble.fakeStallEndedEvent()
                     expect(spy).toEventuallyNot(
-                            haveBeenCalled(withArgs: ["newState": "\(PlayerState.CONVIVA_BUFFERING.rawValue)"])
+                            haveBeenCalled(withArgs: [
+                                CIS_SSDK_PLAYBACK_METRIC_PLAYER_STATE: "\(PlayerState.CONVIVA_BUFFERING.rawValue)"
+                            ])
                         )
                 }
                 it("on stall started / Stall Ended after 0.10 seconds") {
@@ -153,16 +165,20 @@ class PlayerEventsSpec: QuickSpec {
                     DispatchQueue.main.asyncAfter(deadline: .now() + 0.10) {
                         playerDouble.fakeStallEndedEvent()
                         expect(spy).to(
-                            haveBeenCalled(withArgs: ["newState": "\(PlayerState.CONVIVA_BUFFERING.rawValue)"])
+                            haveBeenCalled(withArgs: [
+                                CIS_SSDK_PLAYBACK_METRIC_PLAYER_STATE: "\(PlayerState.CONVIVA_BUFFERING.rawValue)"
+                            ])
                         )
                     }
-                    
+
                 }
                 it("on stall started") {
                     playerDouble.fakeStallStartedEvent()
                     DispatchQueue.main.asyncAfter(deadline: .now() + 0.10) {
                         expect(spy).to(
-                            haveBeenCalled(withArgs: ["newState": "\(PlayerState.CONVIVA_BUFFERING.rawValue)"])
+                            haveBeenCalled(withArgs: [
+                                CIS_SSDK_PLAYBACK_METRIC_PLAYER_STATE: "\(PlayerState.CONVIVA_BUFFERING.rawValue)"
+                            ])
                         )
                     }
                 }
@@ -177,7 +193,9 @@ class PlayerEventsSpec: QuickSpec {
                         _ = TestDouble(aClass: playerDouble, name: "isPlaying", return: true)
                         playerDouble.fakeStallEndedEvent()
                         expect(spy).to(
-                            haveBeenCalled(withArgs: ["newState": "\(PlayerState.CONVIVA_PLAYING.rawValue)"])
+                            haveBeenCalled(withArgs: [
+                                CIS_SSDK_PLAYBACK_METRIC_PLAYER_STATE: "\(PlayerState.CONVIVA_PLAYING.rawValue)"
+                            ])
                         )
                     }
 
@@ -185,7 +203,9 @@ class PlayerEventsSpec: QuickSpec {
                         _ = TestDouble(aClass: playerDouble, name: "isPlaying", return: false)
                         playerDouble.fakeStallEndedEvent()
                         expect(spy).to(
-                            haveBeenCalled(withArgs: ["newState": "\(PlayerState.CONVIVA_PAUSED.rawValue)"])
+                            haveBeenCalled(withArgs: [
+                                CIS_SSDK_PLAYBACK_METRIC_PLAYER_STATE: "\(PlayerState.CONVIVA_PAUSED.rawValue)"
+                            ])
                         )
                     }
                 }
@@ -194,7 +214,7 @@ class PlayerEventsSpec: QuickSpec {
             context("end session") {
                 var spy: Spy!
                 beforeEach {
-                    spy = Spy(aClass: CISClientTestDouble.self, functionName: "cleanupSession")
+                    spy = Spy(aClass: CISVideoAnalyticsTestDouble.self, functionName: "reportPlaybackEnded")
                     playerDouble.fakePlayEvent()
                 }
 
@@ -214,40 +234,36 @@ class PlayerEventsSpec: QuickSpec {
                 }
 
                 it("on playback finished") {
-                    let playbackStateSpy = Spy(aClass: PlayerStateManagerTestDouble.self,
-                                               functionName: "setPlayerState")
+                    let playbackStateSpy = Spy(aClass: CISVideoAnalyticsTestDouble.self,
+                                               functionName: "reportPlaybackMetric")
+
                     playerDouble.fakePlayEvent()
                     playerDouble.fakePlaybackFinishedEvent()
                     expect(spy).to(haveBeenCalled())
                     expect(playbackStateSpy).to(
-                        haveBeenCalled(withArgs: ["newState": "\(PlayerState.CONVIVA_STOPPED.rawValue)"])
+                        haveBeenCalled(withArgs: [
+                            CIS_SSDK_PLAYBACK_METRIC_PLAYER_STATE: "\(PlayerState.CONVIVA_STOPPED.rawValue)"
+                        ])
                     )
                 }
 
                 it("on playlist transition") {
-                    let playbackStateSpy = Spy(aClass: PlayerStateManagerTestDouble.self,
-                                               functionName: "setPlayerState")
+                    let playbackStateSpy = Spy(aClass: CISVideoAnalyticsTestDouble.self,
+                                               functionName: "reportPlaybackMetric")
+
                     playerDouble.fakePlayEvent()
                     playerDouble.fakePlaylistTransitionEvent()
                     expect(spy).to(haveBeenCalled())
                     expect(playbackStateSpy).to(
-                        haveBeenCalled(withArgs: ["newState": "\(PlayerState.CONVIVA_STOPPED.rawValue)"])
+                        haveBeenCalled(withArgs: [
+                            CIS_SSDK_PLAYBACK_METRIC_PLAYER_STATE: "\(PlayerState.CONVIVA_STOPPED.rawValue)"
+                        ])
                     )
                 }
 
                 it("on destroy") {
                     playerDouble.fakeDestroyEvent()
                     expect(spy).to(haveBeenCalled())
-                }
-
-                it("calls end session only if a session is active") {
-                    playerDouble.fakeSourceUnloadedEvent()
-
-                    expect(spy).toEventually(haveBeenCalled())
-                    TestHelper.shared.spyTracker.reset()
-
-                    playerDouble.fakeDestroyEvent()
-                    expect(spy).toEventuallyNot(haveBeenCalled())
                 }
 
                 describe("with on source unloaded / on error workaround") {
@@ -257,8 +273,10 @@ class PlayerEventsSpec: QuickSpec {
 
                         // reset spies to add ability to test against createSession has not been called
                         TestHelper.shared.spyTracker.reset()
-                        let errorSpy = Spy(aClass: CISClientTestDouble.self, functionName: "reportError")
-                        let sessionSpy = Spy(aClass: CISClientTestDouble.self, functionName: "createSession")
+                        let errorSpy = Spy(aClass: CISVideoAnalyticsTestDouble.self,
+                                           functionName: "reportPlaybackError")
+                        let sessionSpy = Spy(aClass: CISVideoAnalyticsTestDouble.self,
+                                             functionName: "reportPlaybackRequested")
 
                         // default sdk error handling is to call unload and this will be triggered first
                         // but we want to track the error event in the same session
@@ -276,8 +294,10 @@ class PlayerEventsSpec: QuickSpec {
 
                         // reset spies to add ability to test against createSession has not been called
                         TestHelper.shared.spyTracker.reset()
-                        let errorSpy = Spy(aClass: CISClientTestDouble.self, functionName: "reportError")
-                        let sessionSpy = Spy(aClass: CISClientTestDouble.self, functionName: "createSession")
+                        let errorSpy = Spy(aClass: CISVideoAnalyticsTestDouble.self,
+                                           functionName: "reportPlaybackError")
+                        let sessionSpy = Spy(aClass: CISVideoAnalyticsTestDouble.self,
+                                             functionName: "reportPlaybackRequested")
 
                         // default sdk error handling is to call unload and this will be triggered first
                         // but we want to track the error event in the same session
@@ -294,42 +314,44 @@ class PlayerEventsSpec: QuickSpec {
             describe("ads") {
                 var spy: Spy!
                 beforeEach {
-                    spy = Spy(aClass: CISClientTestDouble.self, functionName: "adStart")
+                    spy = Spy(aClass: CISVideoAnalyticsTestDouble.self, functionName: "reportAdBreakStarted")
                 }
 
                 context("track preroll ad") {
                     it("with string") {
                         playerDouble.fakeAdStartedEvent(position: "pre")
                         expect(spy).to(
-                            haveBeenCalled(withArgs: ["adPosition": "\(AdPosition.ADPOSITION_PREROLL.rawValue)"])
+                            haveBeenCalled(withArgs: ["adBreakInfo":
+                                                        "\(AdPosition.ADPOSITION_PREROLL.rawValue)"
+                                                     ])
                         )
                     }
 
                     it("with percentage") {
                         playerDouble.fakeAdStartedEvent(position: "0%")
                         expect(spy).to(
-                            haveBeenCalled(withArgs: ["adPosition": "\(AdPosition.ADPOSITION_PREROLL.rawValue)"])
+                            haveBeenCalled(withArgs: ["adBreakInfo": "\(AdPosition.ADPOSITION_PREROLL.rawValue)"])
                         )
                     }
 
                     it("with timestamp") {
                         playerDouble.fakeAdStartedEvent(position: "00:00:00.000")
                         expect(spy).to(
-                            haveBeenCalled(withArgs: ["adPosition": "\(AdPosition.ADPOSITION_PREROLL.rawValue)"])
+                            haveBeenCalled(withArgs: ["adBreakInfo": "\(AdPosition.ADPOSITION_PREROLL.rawValue)"])
                         )
                     }
 
                     it("with invalid position") {
                         playerDouble.fakeAdStartedEvent(position: "start")
                         expect(spy).to(
-                            haveBeenCalled(withArgs: ["adPosition": "\(AdPosition.ADPOSITION_PREROLL.rawValue)"])
+                            haveBeenCalled(withArgs: ["adBreakInfo": "\(AdPosition.ADPOSITION_PREROLL.rawValue)"])
                         )
                     }
 
                     it("without position") {
                         playerDouble.fakeAdStartedEvent(position: nil)
                         expect(spy).to(
-                            haveBeenCalled(withArgs: ["adPosition": "\(AdPosition.ADPOSITION_PREROLL.rawValue)"])
+                            haveBeenCalled(withArgs: ["adBreakInfo": "\(AdPosition.ADPOSITION_PREROLL.rawValue)"])
                         )
                     }
                 }
@@ -338,7 +360,7 @@ class PlayerEventsSpec: QuickSpec {
                     it("with percentage") {
                         playerDouble.fakeAdStartedEvent(position: "10%")
                         expect(spy).to(
-                            haveBeenCalled(withArgs: ["adPosition": "\(AdPosition.ADPOSITION_MIDROLL.rawValue)"])
+                            haveBeenCalled(withArgs: ["adBreakInfo": "\(AdPosition.ADPOSITION_MIDROLL.rawValue)"])
                         )
                     }
 
@@ -346,7 +368,7 @@ class PlayerEventsSpec: QuickSpec {
                         _ = TestDouble(aClass: playerDouble, name: "duration", return: TimeInterval(120))
                         playerDouble.fakeAdStartedEvent(position: "00:01:00.000")
                         expect(spy).to(
-                            haveBeenCalled(withArgs: ["adPosition": "\(AdPosition.ADPOSITION_MIDROLL.rawValue)"])
+                            haveBeenCalled(withArgs: ["adBreakInfo": "\(AdPosition.ADPOSITION_MIDROLL.rawValue)"])
                         )
                     }
                 }
@@ -355,14 +377,14 @@ class PlayerEventsSpec: QuickSpec {
                     it("with string") {
                         playerDouble.fakeAdStartedEvent(position: "post")
                         expect(spy).to(
-                            haveBeenCalled(withArgs: ["adPosition": "\(AdPosition.ADPOSITION_POSTROLL.rawValue)"])
+                            haveBeenCalled(withArgs: ["adBreakInfo": "\(AdPosition.ADPOSITION_POSTROLL.rawValue)"])
                         )
                     }
 
                     it("with percentage") {
                         playerDouble.fakeAdStartedEvent(position: "100%")
                         expect(spy).to(
-                            haveBeenCalled(withArgs: ["adPosition": "\(AdPosition.ADPOSITION_POSTROLL.rawValue)"])
+                            haveBeenCalled(withArgs: ["adBreakInfo": "\(AdPosition.ADPOSITION_POSTROLL.rawValue)"])
                         )
                     }
 
@@ -370,7 +392,7 @@ class PlayerEventsSpec: QuickSpec {
                         _ = TestDouble(aClass: playerDouble, name: "duration", return: TimeInterval(120))
                         playerDouble.fakeAdStartedEvent(position: "00:02:00.000")
                         expect(spy).to(
-                            haveBeenCalled(withArgs: ["adPosition": "\(AdPosition.ADPOSITION_POSTROLL.rawValue)"])
+                            haveBeenCalled(withArgs: ["adBreakInfo": "\(AdPosition.ADPOSITION_POSTROLL.rawValue)"])
                         )
                     }
                 }
@@ -378,7 +400,7 @@ class PlayerEventsSpec: QuickSpec {
                 context("track ad end") {
                     beforeEach {
                         playerDouble.fakePlayEvent()
-                        spy = Spy(aClass: CISClientTestDouble.self, functionName: "adEnd")
+                        spy = Spy(aClass: CISVideoAnalyticsTestDouble.self, functionName: "reportAdBreakEnded")
                     }
 
                     it("on ad skipped") {